+++ hoomd/md/MuellerPlatheFlowGPU.cu
@@ -2,7 +2,6 @@
 // Copyright (c) 2009-2021 The Regents of the University of Michigan
 // This file is part of the HOOMD-blue project, released under the BSD 3-Clause License.
 
-#include "MuellerPlatheFlow.h"
 #include "MuellerPlatheFlowGPU.cuh"
 #include "MuellerPlatheFlowGPU.h"
 #include "hoomd/HOOMDMath.h"
