+++ hoomd/md/MuellerPlatheFlowGPU.cuh
@@ -1,7 +1,7 @@
 // Copyright (c) 2009-2021 The Regents of the University of Michigan
 // This file is part of the HOOMD-blue project, released under the BSD 3-Clause License.
 
-#include "MuellerPlatheFlow.h"
+#include "MuellerPlatheFlowEnum.h"
 #include "hoomd/HOOMDMath.h"
 #include "hoomd/ParticleData.cuh"
